test: Fix integration test isolation
